Split equirectangular images into two textures if too big.
This allows for equirectangular images on a given device to be twice as big as
they were previously, so (almost) all devices now support equirectangular
images up to 8192px across.

This technique could be extended to double the maximum size again (using eight
textures), but images bigger than 8192px should be converted to the
multi-resolution format anyway, so there's little point in implementing this
extension.